testing ARAGBT2 Mega model fine-tuned on Ibn Taymiyyah's texts
This notebook demonstrates the application of the ARGBT2 Mega model from Hugging Face, fine-tuned on the texts of Sheikh Al-Islam Ibn Taymiyyah. The experiment uses LoRa configuration and quantizes the model to 4-bit MF4 to assess its performance in generating text similar to Sheikh Al-Islam's style. The notebook is intended for evaluating model behavior and provides insights into its potential for working with classical Arabic texts